Re #115: Attempt to fix linux build warnings
git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@1425 25a7314f-c4c8-4314-9936-a25f8480a772

+++ SlicerRtCommon/SlicerRtCommon.cxx
@@ -286,7 +286,7 @@ void SlicerRtCommon::StretchDiscreteColorTable(vtkMRMLColorTableNode* inputDiscr
     return;
   }
 
-  if (inputDiscreteColorTable->GetNumberOfColors() >= numberOfColors)
+  if (inputDiscreteColorTable->GetNumberOfColors() >= (int)numberOfColors)
   {
     vtkErrorWithObjectMacro(inputDiscreteColorTable, "SlicerRtCommon::StretchDiscreteColorTable: Input discrete color table should have less colors than the specified number of colors (" << inputDiscreteColorTable->GetNumberOfColors() << " < " << numberOfColors << ")");
     return;
